Fix: Taking mean over the wrong dimension
The tree explainer outputs log odds for for each class. By taking a mean of the 0th dim, we are reducing this object to just 2 values - one of for each class in a binary classification problem.

Instead, we need to take the mean of the 2nd dimension, which after the previous mean is the 2 classes.